First pass at a run tree viewer Qt widget
This example uses example data to show a single run in a tree view using
the new API. It lazily populates the tree as items are expanded, and
takes a somewhat opinionated view on how some elements are exposed.
There are some shortcomings with expansion of lists containing dicts
that need further refinement.