Make Search tabs not closable.
Search tabs can be added or removed from the model (e.g. console).
Currently, they can be removed (closed) from the UI but not added. This
seems bad, as you could accidentally close a tab and have no way to get
it back. For now, I think it's better to make the tabs not closable from
the UI.