PERF: Convert zmq dispatcher LOADING_LATENCY to msec
Had been getting high CPU usage and balooning memory usage ultimately causing the OS to kill my application. Changing this one number fixed that. It is documented as an integer in msec, so a value of 0.01 does not make sense

Additionally singleShot is a staticmethod and so does not need to be called on an instance of QTimer